refactor(components): 顶层组件按领域拆分到目录
components/ 下扁平堆叠的 30 个组件按领域归入子目录：
- common/   基础公共 UI（Avatar / ConfirmModal / Loading / icons / Mermaid 等）
- layout/   应用骨架（MainPane / Sidebar / StatusBar / TitleBar）
- pr/ · diff/ · comments/ · drafts/ · settings/   各业务领域

纯文件位置调整 + import 路径改写（git 识别为 rename），无逻辑/界面变更。
