+++ apps/desktop/src/renderer/src/components/chat/components/FindingCard.tsx
@@ -4,8 +4,8 @@ import ReactMarkdown from 'react-markdown';
 import remarkBreaks from 'remark-breaks';
 import remarkGfm from 'remark-gfm';
 import type { Finding, PrDocSectionKey, ReviewDraft } from '@meebox/shared';
-import { ChevronIcon } from '../../icons';
-import { mermaidComponents, walkthroughMdComponents } from '../../markdownMermaid';
+import { ChevronIcon } from '../../common/icons';
+import { mermaidComponents, walkthroughMdComponents } from '../../common/markdownMermaid';
 import { REMOTE_REHYPE_PLUGINS } from '../../../markdown';
 import { translatePrAgentLabels } from '../../../utils/translate-pr-agent';
 import {
