+++ apps/desktop/src/renderer/src/components/common/LlmProviderIcon.tsx
