+++ apps/desktop/src/renderer/src/components/common/MermaidDiagram.tsx
