+++ apps/desktop/src/renderer/src/components/common/PlatformIcon.tsx
