+++ apps/desktop/src/renderer/src/components/common/icons.tsx
