+++ apps/desktop/src/renderer/src/components/common/markdownMermaid.tsx
