+++ apps/desktop/src/renderer/src/components/layout/MainPane.tsx
@@ -6,17 +6,17 @@ import type {
   ReviewerStatus,
   StoredPullRequest,
 } from '@meebox/shared';
-import { invoke } from '../api';
-import { useDraftsForPr } from '../stores/drafts-store';
-import { CommentsPanel } from './CommentsPanel';
-import { CommitsPanel } from './CommitsPanel';
+import { invoke } from '../../api';
+import { useDraftsForPr } from '../../stores/drafts-store';
+import { CommentsPanel } from '../comments/CommentsPanel';
+import { CommitsPanel } from '../pr/CommitsPanel';
 // Monaco 编辑器（~10MB）懒加载：只有真正切到 Diff tab 才拉取 DiffView chunk，
 // 不阻塞窗口首帧 / PR 列表 / 首启向导。
-const DiffView = lazy(() => import('./DiffView').then((m) => ({ default: m.DiffView })));
-import { DraftsPanel } from './DraftsPanel';
-import { PaneLoading } from './Loading';
-import { PrInfoView } from './PrInfoView';
-import { PublishReviewModal } from './PublishReviewModal';
+const DiffView = lazy(() => import('../diff/DiffView').then((m) => ({ default: m.DiffView })));
+import { DraftsPanel } from '../drafts/DraftsPanel';
+import { PaneLoading } from '../common/Loading';
+import { PrInfoView } from '../pr/PrInfoView';
+import { PublishReviewModal } from '../drafts/PublishReviewModal';
 import {
   ApproveIcon,
   GlobeIcon,
@@ -24,7 +24,7 @@ import {
   PersonIcon,
   PullRequestIcon,
   WhitespaceIcon,
-} from './icons';
+} from '../common/icons';
 
 interface MainPaneProps {
   pr: StoredPullRequest | null;
