+++ apps/desktop/src/renderer/src/components/onboarding/OnboardingWizard.tsx
@@ -8,8 +8,8 @@ import {
   fromConnDraft,
   type ConnDraft,
   type ConnEntry,
-} from '../ConnectionForm';
-import { newProfileId } from '../LlmProfileForm';
+} from '../settings/ConnectionForm';
+import { newProfileId } from '../settings/LlmProfileForm';
 import { WelcomeStep } from './steps/WelcomeStep';
 import { PlatformStep } from './steps/PlatformStep';
 import { LlmStep } from './steps/LlmStep';
