+++ apps/desktop/src/renderer/src/components/settings/ConnectionForm.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { TFunction } from 'i18next';
 import { GITHUB_DOTCOM_API_BASE, GITLAB_DOTCOM_API_BASE, type Config } from '@meebox/shared';
-import { invoke } from '../api';
-import { EyeIcon, EyeOffIcon } from './icons';
+import { invoke } from '../../api';
+import { EyeIcon, EyeOffIcon } from '../common/icons';
 
 // 连接编辑用的扁平草稿（Connection 是嵌套的 auth/clone，拍平后表单好写），存盘前还原。
 // 设置页 ConnectionEditorModal 与首启向导 PlatformStep 共用同一份草稿形状 + 表单。
